Refactoring: API and internal modules structure changed
