+++ samples/unstructured_data/update_model.py
@@ -2,5 +2,4 @@ from st_library import Library
 
 st_lib = Library()
 st_lib.set_token('Bearer token')
-print(st_lib.upload_file("19a29b9b-bea2-40fb-89c4-555bba829539","RandomForest_train.rds","../data/unstructured_data/"))
-
+print(st_lib.unstruct_data.upload_file("19a29b9b-bea2-40fb-89c4-555bba829539","RandomForest_train.rds","../data/unstructured_data/"))
