+++ st_library/core/dataprovider/structured_data/structured_data.py
@@ -1,194 +1,30 @@
-# Copyright 2017 Shortest Track Company. All rights reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
-# in compliance with the License. You may obtain a copy of the License at
-#
-# http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software distributed under the License
-# is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
-# or implied. See the License for the specific language governing permissions and limitations under
-# the License.
-
-"""Implements BigQuery HTTP API wrapper."""
-
-import time
-
-from st_library.utils.helpers.store import Store
-from st_library.utils.api_client import http, endpoints
+from st_library.core.dataprovider.structured_data.table import Table
 
 
 class StructuredData(object):
-    """A helper class to issue BigQuery HTTP requests."""
+    def __init__(self):
+        self.Table = Table
 
-    def tables_get(self, matrices_id):
-        """Issues a request to retrieve information about a table.
-
-        Args:
-          matrices_id: .
-        Returns:
-          A parsed result object.
-        Raises:
-          Exception if there is an error performing the operation.
+    def read_matrix(self, matricesid, datasetsid, tablename):
         """
-        return http.Http.request(endpoints.matrices_path(matrices_id))
-
-    def tables_get_parameter(self):
-        """Issues a request to retrieve the parameter of the script execution configuration.
-
-        Args:
-        Returns:
-          A parsed result object.
-        Raises:
-          Exception if there is an error performing the operation.
+        Retrieves the table data.
+        :return dataframe:
         """
-        return http.Http.request(endpoints.sec_get_detail_path(Store.config_id))
+        tbl = self.Table(matricesid, datasetsid, tablename)
+        return (tbl.to_dataframe())
 
-    def tabledata_list(self, config_related, datasetsid, table_name, start_index=None, max_results=None, page_token=None):
-        """ Retrieves the contents of a table.
-
-        Args:
-          table_name: the name of the table as a tuple of components.
-          start_index: the index of the row at which to start retrieval.
-          max_results: an optional maximum number of rows to retrieve.
-          page_token: an optional token to continue the retrieval.
-        Returns:
-          A parsed result object.
-        Raises:
-          Exception if there is an error performing the operation.
+    def read_sec_matrix(self, matricesid, datasetsid, tablename):
         """
-        if config_related:
-            url = endpoints.sec_matrices_get_path(Store.config_id, datasetsid, table_name)
-        else:
-            url = endpoints.matrices_get_path(datasetsid, table_name)
-        args = {}
-        if start_index:
-            args['startIndex'] = start_index
-        if max_results:
-            args['maxResults'] = max_results
-        if page_token is not None:
-            args['pageToken'] = page_token
-        return http.Http.request(url, args=args)
-
-    def tabledata_post(self, datasetsid, table_name, file_name):
-        """ Insert the contents of a table.
-
-        Args:
-          file_name: the name of the table as a tuple of components.
-        Returns:
-          A result object.
-        Raises:
-          Exception if there is an error performing the operation.
+        Retrieves the table data.
+        :return dataframe:
         """
+        tbl = self.Table(matricesid, datasetsid, tablename, config_related=True)
+        return (tbl.to_dataframe())
 
-        url = endpoints.matrices_upload_path(datasetsid, table_name)
-
-        filepath1 = r"/home/st/workspace/st-dataprovider-python/datalab/structured_data/metadata.json"
-        filepath2 = file_name
-        boundary = '----------%s' % hex(int(time.time() * 1000))
-        data = []
-        data.append('--%s' % boundary)
-        fr1 = open(filepath1, 'rb')
-        data.append('Content-Disposition: form-data; name="%s"; filename="metadata.json"' % 'metadata')
-        data.append('Content-Type: %s\r\n' % 'application/json')
-        data.append(fr1.read())
-        fr1.close()
-        data.append('--%s' % boundary)
-        fr2 = open(filepath2, 'rb')
-        data.append('Content-Disposition: form-data; name="%s"; filename="data.csv"' % 'file')
-        data.append('Content-Type: %s\r\n' % 'application/vnd.ms-excel')
-        data.append(fr2.read())
-        fr2.close()
-        data.append('--%s--\r\n' % boundary)
-
-        http_body = '\r\n'.join(data)
-
-        headers = {}
-        headers['Content-Type'] = 'multipart/form-data; boundary=%s' % boundary
-
-        return http.Http.request(url=url, data=http_body, headers=headers, method='POST')
-
-    def insert_sec_data(self, datasetsid, table_name, json_data):
-        """ Insert streams data into matrix.
-
-        Args:
-          file_name: the name of the table as a tuple of components.
-        Returns:
-          A result object.
-        Raises:
-          Exception if there is an error performing the operation.
+    def get_parameter(self):
         """
-        url = endpoints.sec_matrices_insert_path(Store.config_id, datasetsid, table_name)
-        return http.Http.request(url=url, data=json_data, method='POST')
-
-    def insert_batch_sec_data(self, datasetsid, table_name, file_path, file_name):
-        """ Insert streams data into matrix.
-
-        Args:
-          file_name: the name of the table as a tuple of components.
-        Returns:
-          A result object.
-        Raises:
-          Exception if there is an error performing the operation.
+        Retrieves the table data.
+        :return dataframe:
         """
-        url = endpoints.sec_matrices_batch_insert_path(Store.config_id, datasetsid, table_name)
-
-        metadata_file = '/home/st/workspace/dataprovider-py/samples/data/structured_data/metadata_sec.json'
-        data_file = file_path + file_name
-        boundary = '----------%s' % hex(int(time.time() * 1000))
-        data = []
-        data.append('--%s' % boundary)
-        fr1 = open(metadata_file, 'rb')
-        data.append('Content-Disposition: form-data; name="%s"; filename="metadata_sec.json"' % 'metadata')
-        data.append('Content-Type: %s\r\n' % 'application/json')
-        data.append(fr1.read())
-        fr1.close()
-        data.append('--%s' % boundary)
-        fr2 = open(data_file, 'rb')
-        data.append('Content-Disposition: form-data; name="%s"; filename="data.csv"' % 'data')
-        data.append('Content-Type: %s\r\n' % 'application/vnd.ms-excel')
-        data.append(fr2.read())
-        fr2.close()
-        data.append('--%s--\r\n' % boundary)
-
-        http_body = '\r\n'.join(data)
-
-        headers = {}
-        headers['Content-Type'] = 'multipart/form-data; boundary=%s' % boundary
-
-        return http.Http.request(url=url, data=http_body, headers=headers, method='POST')
-
-        # if sys.version > '3':
-        #   from requests_toolbelt.multipart.encoder import MultipartEncoder
-        #
-        #   multipart_data = MultipartEncoder(
-        #     fields={
-        #       'metadata':('metadata_sec.json',
-        #                open('/home/st/workspace/dataprovider-py/samples/data/structured_data/metadata_sec.json',
-        #                     'rb'), 'text/plain'),
-        #       # a file upload field
-        #       'data': (file_name,
-        #                open(file_path+file_name,
-        #                     'rb'), 'text/plain')
-        #     }
-        #   )
-        #   headers = {}
-        #   headers['Content-Type'] = multipart_data.content_type
-        #
-        #   return _http.Http.request(url=url, data=multipart_data, headers=headers, method='POST')
-        # else:
-        #   from poster.encode import multipart_encode
-        #   from poster.streaminghttp import register_openers
-        #
-        #   # Register the streaming http handlers with urllib2
-        #   register_openers()
-        #
-        #   # Start the multipart/form-data encoding of the file
-        #   # "file" is the name of the parameter, which is normally set
-        #   # via the "name" parameter of the HTML <input> tag.
-        #   # headers contains the necessary Content-Type and Content-Length
-        #   # datagen is a generator object that yields the encoded parameters
-        #   datagen, headers = multipart_encode(
-        #     {"metadata": open('/home/st/workspace/dataprovider-py/samples/data/structured_data/metadata_sec.json','rb'),"data": open(file_path+file_name,'rb')})
-        #
-        #   return _http.Http.request(url=url, data=datagen, headers=headers, method='POST')
+        tbl = self.Table()
+        return tbl.get_parameter_data()
