+++ st_library/core/dataprovider/unstructured_data/__init__.py
@@ -9,7 +9,3 @@
 # is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
 # or implied. See the License for the specific language governing permissions and limitations under
 # the License.
-
-from .item import Item
-
-__all__ = ['Item']
