+++ st_library/library.py
@@ -9,9 +9,8 @@
 # is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
 # or implied. See the License for the specific language governing permissions and limitations under
 # the License.
-from st_library.core.dataprovider import Item
-from st_library.core.dataprovider import Table
-from st_library.core.dataprovider.dataprovider import DataProvider
+from st_library.core.dataprovider.structured_data.structured_data import StructuredData
+from st_library.core.dataprovider.unstructured_data.unstructured_data import UnstructuredData
 from st_library.core.logger.logger import Logger
 from st_library.core.metadata.metadata import Metadata
 from st_library.utils.generics.singleton import Singleton
@@ -27,9 +26,10 @@ class Library(Singleton):
     """
 
     def __init__(self):
+        self.struct_data = StructuredData()
+        self.unstruct_data = UnstructuredData()
         self.metadata = Metadata()
         self.logger = Logger()
-        self.data_provider = DataProvider()
 
     @property
     def config_id(self):
@@ -62,51 +62,3 @@ class Library(Singleton):
         :return:
         """
         return Store.performance_id
-
-    def read_matrix(self, matricesid, datasetsid, tablename):
-        """
-        Retrieves the table data.
-        :return dataframe:
-        """
-        tbl = Table(matricesid, datasetsid, tablename)
-        return (tbl.to_dataframe())
-
-    def read_sec_matrix(self, matricesid, datasetsid, tablename):
-        """
-        Retrieves the table data.
-        :return dataframe:
-        """
-        tbl = Table(matricesid, datasetsid, tablename, config_related=True)
-        return (tbl.to_dataframe())
-
-    def get_parameter(self):
-        """
-        Retrieves the table data.
-        :return dataframe:
-        """
-        tbl = Table()
-        return tbl.get_parameter_data()
-
-    def download_file(self, datasetsid, filename):
-        """
-        Download files to disk.
-        :return local file path:
-        """
-        item = Item(datasetsid=datasetsid)
-        return item.download_file(filename)
-
-    def upload_file(self, datasetsid, filename, filepath):
-        """
-        Download dataset files to disk.
-        :return local file path:
-        """
-        item = Item(datasetsid=datasetsid)
-        return item.upload_file(filename, filepath)
-
-    def delete_file(self, datasetsid, filename):
-        """
-        Download files to disk.
-        :return local file path:
-        """
-        item = Item(datasetsid=datasetsid)
-        return item.delete_file(filename)
